Move tests to src dir
