+++ src/tests/timer-reducer.test.js
@@ -1,7 +1,7 @@
 import moment       from 'moment';
-import TimerEvents  from '../src/js/enums/TimerEvents';
-import reducer      from '../src/js/reducers/TimerReducer';
-import TimeTracker  from '../src/js/containers/TimeTracker'
+import TimerEvents  from '../js/enums/TimerEvents';
+import reducer      from '../js/reducers/TimerReducer';
+import TimeTracker  from '../js/containers/TimeTracker'
 
 // Defaults
 const defaultState = {timers: []};
